+++ CHANGELOG.md
@@ -3,8 +3,13 @@
 ## Unreleased
 - unicast emulation.
 - sound.
-- restriction (middleware).
-- logout
+
+
+## v0.4.22 / 2024-06-24
+- restriction inside index.
+- logout. `#(assoc % session {})`.
+- moved `/resources/public/index.html` to `/resources/index.html`.
+- redirect / to /login? -> moved `/login` as `/`.
 
 ## v0.3.18 / 2024-06-24
 - auth.
@@ -12,7 +17,6 @@
 - hk-server is not good against ring-example server.
   differ websocket API, jetty and http-kit?
 - hk-client, too. need cachire to decode? (hc/get url {:as :json}) is fine.
-
 
 ## v0.2.12 / 2024-06-24
 - moved default middleware to "/chat".